requirements.txt
faster-whisper>=0.10.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
rumps>=0.4.0
pynput>=1.7.6
PyYAML>=6.0
